Refactor index.js to keep it DRY. Related #10.
