+++ index.html
@@ -53,6 +53,7 @@
     <div class="game__messages"></div>
 
     <script type="text/javascript" src="./public/index.js"></script>
+    <script type="text/javascript" src="./public/index.js"></script>
     <script type="text/javascript" src="./public/game.js"></script>
   </body>
 </html>
