+++ tests/makeWinningMove.test.js
@@ -1,5 +1,5 @@
 var test = require('tape');
-var game = require('../game.js');
+var game = require('../public/game.js');
 
 test('Testing that makeWinningMove() returns the correct next position', function(t){
   var board = [' ', 'O', 'O', '.', '.', '.', '.', '.', '.'];
